*Продолжение глобализации безопасности* Шкала остроты - только численная Выбор остроты соуса

+++ Program.cs
@@ -9,8 +9,8 @@ namespace Шава_мейкер_2._0
     static class Data
     {
         public static double Summ { get; set; } //Глобальная переменная для подсчета стоимости шавы
-        public static double Ost { get; set; } //Глобальная переменная для шкалы остраты
-        public static double Callories { get; set; } //Глобальная переменная для подсчета калорийности шавы
+        public static int Ost { get; set; } //Глобальная переменная для шкалы остраты
+        public static int Callories { get; set; } //Глобальная переменная для подсчета калорийности шавы
 }
     static class Program
     {
